refactor(tools): remove touch tool; allow empty write content
